Allow collection of system profile facts (part I)
We would like to collect "system profile" facts. These are different than
canonical facts since they are not used to identify a machine.

This commit adds "system_profile" to the facts hash and sends it into inventory
service. For now, the additional facts are placed under the `system_profile`
namespace.

We also add an additional util to get the system profile facts, which relies on
insights-core rules and parsers.

This commit does not include the full set of facts we wish to fetch. However,
it is a sizable set (>50%) and contains larger fact sets like available systemd
services (installed packages is not included). Another PR will add the
remainder of the facts.