pull cloud_provider for system profile
A recent insights-core commit added the `CloudProvider` combiner,
which will output a system's cloud provider ("aws", "azure", etc).

This commit adds the combiner to PUP and puts the field into the
system profile as `cloud_provider`. Note that this will not populate
in inventory service until the field exists there.